requirements.txt
aiohttp==3.7.3
aiohttp-json-rpc==0.13.3
aiosqlite==0.16.1
appdirs==1.4.4
apsw-wheels==3.34.1.post1
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
balls==0.0.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
chat-exporter==1.4.1
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
contextlib2==0.6.0.post1
discord-components==0.5.3
discord-sentry-reporting==2.0.4
discord.py==1.7.2
distro==1.5.0
emoji==1.2.0
Flask==2.0.1
fuzzywuzzy==0.18.0
google-auth==1.32.0
google-auth-oauthlib==0.4.4
grapheme==0.6.0
gspread==3.7.0
httplib2==0.19.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
Markdown==3.3.3
MarkupSafe==2.0.1
multidict==5.1.0
oauth2client==4.1.3
oauthlib==3.1.1
peewee==3.14.4
Pillow==8.2.0
pkg-resources==0.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pygicord==0.5.0
Pygments==2.7.4
PyMySQL==1.0.2
PyNaCl==1.3.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.1
python-Levenshtein-wheels==0.13.2
pytz==2021.1
PyYAML==5.4.1
Red-DiscordBot==3.4.10
Red-Lavalink==0.8.1
requests==2.25.1
requests-oauthlib==1.3.0
rich==9.9.0
rsa==4.7.2
schema==0.7.4
sentry-sdk==1.1.0
six==1.15.0
tqdm==4.56.2
typing-extensions==3.7.4.3
urllib3==1.26.5
uvloop==0.15.0
Werkzeug==2.0.1
yarl==1.6.3
youtube-dl==2021.6.6
